Organize: move internals to impl header, remove redundant usings
