+++ cloud/filestore/libs/vfs_fuse/write_back_cache/calculate_data_parts_to_read_ut.cpp
@@ -1,4 +1,4 @@
-#include "write_back_cache.h"
+#include "write_back_cache_impl.h"
 
 #include <cloud/storage/core/libs/diagnostics/logging.h>
