+++ cloud/filestore/libs/vfs_fuse/write_back_cache/session_sequencer.cpp
@@ -1,9 +1,5 @@
 #include "session_sequencer.h"
 
-#include <util/generic/algorithm.h>
-#include <util/generic/hash.h>
-#include <util/system/mutex.h>
-
 namespace NCloud::NFileStore::NFuse {
 
 using namespace NThreading;
