+++ cloud/filestore/libs/vfs_fuse/write_back_cache/write_back_cache.cpp
@@ -1,4 +1,4 @@
-#include "write_back_cache.h"
+#include "write_back_cache_impl.h"
 
 #include "session_sequencer.h"
 
@@ -10,14 +10,11 @@
 
 #include <util/generic/hash_set.h>
 #include <util/generic/intrlist.h>
-#include <util/generic/mem_copy.h>
-#include <util/generic/strbuf.h>
 #include <util/generic/vector.h>
 #include <util/stream/mem.h>
 #include <util/system/mutex.h>
 
 #include <algorithm>
-#include <atomic>
 
 namespace NCloud::NFileStore::NFuse {
